# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Aplikacija za izložbe fotografija",
        "description": "Swagger dokumentacija Laravel API-ja",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalno (Docker)"
        }
    ],
    "paths": {
...
        "/api/korisnici": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Lista korisnika (admin)",
                "operationId": "5061f6d16eb244e5b6757a44af053383",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Kreiranje korisnika (admin)",
                "operationId": "b09c5a6427c0ae9a7ce73389ef77acf3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "prezime",
                                    "email",
                                    "lozinka",
                                    "uloga"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Ana"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Anić"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana@test.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "example": "ana12345"
                                    },
                                    "uloga": {
                                        "type": "string",
                                        "example": "fotograf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kreirano"
                    },
                    "403": {
                        "description": "Zabranjeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/korisnici/{id}": {
            "put": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Izmena korisnika (admin)",
                "operationId": "ce20b76969d640b2ad7066574cc652e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Ana"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Anić"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana@test.com"
                                    },
                                    "lozinka": {
                                        "type": "string",
                                        "example": "nova12345"
                                    },
                                    "uloga": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Brisanje korisnika (admin)",
                "operationId": "fb63425b86922025564614af57838826",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisan"
                    },
                    "404": {
                        "description": "Nije pronađen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/korisnici/{id}/izlozbe": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Izložbe na koje je korisnik prijavljen",
                "operationId": "078c3385ba5cfdc58f0ace780d04b816",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Nije pronađen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Unesi token u formatu: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}